fix: move entities to database folder and entities into entities
