+++ src/database/seeding/seeding.module.ts
@@ -2,7 +2,7 @@ import { Module } from '@nestjs/common';
 import { SeedingService } from './seeding.service';
 import { SeedingController } from './seeding.controller';
 import { TypeOrmModule } from '@nestjs/typeorm';
-import { User } from '../../modules/users/user.entity';
+import { User } from '../../modules/users/entities/user.entity';
 @Module({
   imports: [TypeOrmModule.forFeature([User])],
   providers: [SeedingService],
